// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/door-link-snapshot-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DoorLinkSnapshotAyahDto } from '../models/door-link-snapshot-ayah-dto';
import { DoorLinkSnapshotRecordDto } from '../models/door-link-snapshot-record-dto';
export interface DoorLinkSnapshotDto {
  ayahs: Array<DoorLinkSnapshotAyahDto>;
  doorId: number;
  doorVersion: number;
  linkingDataRevision: number;
  records: Array<DoorLinkSnapshotRecordDto>;
}